app1 app2 成功改写MVC并且修复app1小数不显示问题

+++ src/app1/app1.js
@@ -8,7 +8,7 @@ const init = (el) => {
         eventBus: eventBus,
         el: $(el),
         data: {
-            n: parseInt(localStorage.getItem("n")) || 100
+            n: parseFloat(localStorage.getItem("n")) || 100
         },
         updated(data) {
             Object.assign(this.data, data)
